Add standard hyphenation support.
