+++ src/SixLabors.Fonts/LogicalTextLine.cs
@@ -16,20 +16,23 @@ internal readonly struct LogicalTextLine
     /// <param name="textLine">The composed logical text line.</param>
     /// <param name="lineBreaks">The collected line break opportunities.</param>
     /// <param name="wordSegments">The collected word-boundary segment runs.</param>
+    /// <param name="hyphenationMarkers">The visible hyphenation markers created for soft hyphen entries.</param>
     public LogicalTextLine(
-        TextLayout.TextLine textLine,
+        TextLine textLine,
         List<LineBreak> lineBreaks,
-        List<WordSegmentRun> wordSegments)
+        List<WordSegmentRun> wordSegments,
+        List<GlyphLayoutData> hyphenationMarkers)
     {
         this.TextLine = textLine;
         this.LineBreaks = lineBreaks;
         this.WordSegments = wordSegments;
+        this.HyphenationMarkers = hyphenationMarkers;
     }
 
     /// <summary>
     /// Gets the composed logical text line.
     /// </summary>
-    public TextLayout.TextLine TextLine { get; }
+    public TextLine TextLine { get; }
 
     /// <summary>
     /// Gets the collected line break opportunities.
@@ -40,4 +43,9 @@ internal readonly struct LogicalTextLine
     /// Gets the collected word-boundary segment runs.
     /// </summary>
     public List<WordSegmentRun> WordSegments { get; }
+
+    /// <summary>
+    /// Gets the visible hyphenation markers created for soft hyphen entries.
+    /// </summary>
+    public List<GlyphLayoutData> HyphenationMarkers { get; }
 }
