+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsBidiSoftHyphenMarker_109.48-HorizontalBottomTop.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7b2b4d5f310eecfae93f9bff0531ab2a7741a24acfe95de30638974467d3d5a3
+size 7486
