+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsBidiSoftHyphenMarker_109.48-HorizontalTopBottom.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:20aba76bd33e13e45d5822ab8bc014624393ce507ef4790ef737b1d15fa77ead
+size 7492
