+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsBidiSoftHyphenMarker_109.48-VerticalMixedLeftRight.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c1712941c5188637c7339ef8765c1322ffbd71d88af2d0a098ae0daed1100c95
+size 7247
