+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsBidiSoftHyphenMarker_109.48-VerticalMixedRightLeft.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:04863a6bc815ec8ad71a551e0f9169031854b84dc42cb366b97da5ba00766bf6
+size 7191
