+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsBidiSoftHyphenMarker_327.76-VerticalLeftRight.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b01ed4070f9d18b20a885ab1eee5b787b370eef2168963412569d06dc324dfaa
+size 9236
