+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsBidiSoftHyphenMarker_327.76-VerticalRightLeft.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0ead5ddc41c56cf3017dd457cec1d078301cfeb1c863a71b0c159f00add85b4d
+size 9084
