+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsSelectedSoftHyphenMarker_246.16-VerticalLeftRight.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:89ecaf4da858091b7a07aa9a52a6ba7cec9977b80ac085cf5a98c68ad347eab0
+size 3270
