+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsSelectedSoftHyphenMarker_246.16-VerticalRightLeft.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:71e9fbc70d5f379c6cbf6a39891939c952ed9c382f868b894d451c4aab121457
+size 3275
