+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsSelectedSoftHyphenMarker_89.17-HorizontalBottomTop.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c961b52c626acf4beef744e7606f725f2da9c701c16ceccf8880d23cebecf7c6
+size 3239
