+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsSelectedSoftHyphenMarker_89.17-HorizontalTopBottom.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f3f943d2c4b1b070787a3814adff064eca0ae87d2817271ed8ef1e913d80ec3a
+size 3256
