+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsSelectedSoftHyphenMarker_89.17-VerticalMixedLeftRight.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a522dba8cf3fca4b40c411a986a035da73412456c37b1f88a496644b95567e26
+size 3038
