+++ tests/Images/ReferenceOutput/TextHyphenation_Custom_DrawsSelectedSoftHyphenMarker_89.17-VerticalMixedRightLeft.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:3170a198b6f1749865a6421f84248e6deb5f2f035f9cf0121106558e46ae877e
+size 3041
