+++ tests/Images/ReferenceOutput/TextHyphenation_None_DoesNotDrawFallbackSoftHyphenMarker_100.48-HorizontalBottomTop.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7e8fb48aee2195182d786cdc5b08c5270a377ecccac760c068b54ec01e1878db
+size 4285
