+++ tests/Images/ReferenceOutput/TextHyphenation_None_DoesNotDrawFallbackSoftHyphenMarker_100.48-HorizontalTopBottom.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:355adc5176c29b1e72bd042adc5e440516599c57ef742794037ca3e54e9f9556
+size 4276
