+++ tests/Images/ReferenceOutput/TextHyphenation_None_DoesNotDrawFallbackSoftHyphenMarker_100.48-VerticalMixedLeftRight.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:83b64489947de6a2279c7fb8ab4fbf70217252245d6b596a3573e7495368e4b4
+size 4041
