+++ tests/Images/ReferenceOutput/TextHyphenation_None_DoesNotDrawFallbackSoftHyphenMarker_100.48-VerticalMixedRightLeft.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:51e674f3bc57adff8458bd16bddaad5259594c0503165dadac6cc2b6a244286e
+size 4009
