+++ tests/Images/ReferenceOutput/TextHyphenation_None_DoesNotDrawFallbackSoftHyphenMarker_326.19998-VerticalLeftRight.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:60a91d50c803a82a521d3443688e1a47c6900976396e1496d7bdb3edf7b3f1fb
+size 7966
