+++ tests/Images/ReferenceOutput/TextHyphenation_None_DoesNotDrawFallbackSoftHyphenMarker_326.19998-VerticalRightLeft.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:77ce2e45a1ae9f5c4e38227f5c09de6d23e6cbaa4807e220c2c56465cdb5f9ad
+size 8109
