+++ tests/Images/ReferenceOutput/TextHyphenation_Standard_DrawsFallbackSoftHyphenMarker_101.61001-HorizontalBottomTop.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6b14fb9547656d10a0686379af6ab33ef6133ad99ee09ba8a7facc810ca8ef15
+size 4131
