+++ tests/Images/ReferenceOutput/TextHyphenation_Standard_DrawsFallbackSoftHyphenMarker_101.61001-HorizontalTopBottom.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:d77c56d7ea96cd93034761b14122a19ea76240d02f10c44b6b2272ac39843c2d
+size 4132
