+++ tests/Images/ReferenceOutput/TextHyphenation_Standard_DrawsFallbackSoftHyphenMarker_101.61001-VerticalMixedLeftRight.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f82fd5cf59b10d73256adcfddc36a3d5da64b33123a6a341632e71def458600a
+size 4022
