+++ tests/Images/ReferenceOutput/TextHyphenation_Standard_DrawsFallbackSoftHyphenMarker_101.61001-VerticalMixedRightLeft.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:11d9bf37b0c3e878684d95db8487af79dbce24aa373837855fed2bfcfd86e852
+size 6353
