+++ tests/Images/ReferenceOutput/TextHyphenation_Standard_DrawsFallbackSoftHyphenMarker_326.76-VerticalLeftRight.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:e87d2e735f322fd9843b5624566af47d38fe344979f355c356d24a0f67184e8e
+size 4402
