+++ tests/Images/ReferenceOutput/TextHyphenation_Standard_DrawsFallbackSoftHyphenMarker_326.76-VerticalRightLeft.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:f8554605300873c2a52a90a7bc5dbbb1d8c8c918fb84bb43919757e9d59d0092
+size 4364
